requirements.txt
feedparser==6.0.8
Flask==2.0.2
numpy==1.21.4
scikit-learn==1.0.1
sqlitedict==1.7.0
requests==2.28.1
sendgrid==6.9.7
google-cloud-storage==2.5.0
